zora-signal-bot/pyproject.toml
[project]
name = "zora-signal-bot"
version = "0.1.0"
description = "Telegram bot that monitors X/Twitter accounts and detects Zora coin trading opportunities"
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Telegram
    "python-telegram-bot[webhooks]>=21.6",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.29.0",
    "alembic>=1.13.3",
    # Cache / queue
    "redis>=5.1.0",
    "celery[redis]>=5.4.0",
    # Validation / config
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    # HTTP client
    "httpx>=0.27.2",
    # Logging
    "structlog>=24.4.0",
    # Utilities
    "python-dateutil>=2.9.0",
    "tenacity>=9.0.0",
    "orjson>=3.10.7",
    "eth-account>=0.11.0",
    "eth-abi>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.2",
    "factory-boy>=3.3.1",
    "respx>=0.21.1",
    "ruff>=0.6.9",
    "mypy>=1.11.2",
    "types-python-dateutil>=2.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing -v"

[tool.coverage.run]
source = ["app"]
omit = ["*/tests/*", "*/migrations/*"]
